chore(lsadf-api): changing some packages names in lsadf-api module
