--- lsadf_api/src/main/java/com/lsadf/application/controller/game/characteristics/CharacteristicsController.java
@@ -1,120 +1,0 @@
-/*
- * Copyright © 2024-2025 LSDAF
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.lsadf.application.controller.game.characteristics;
-
-import static com.lsadf.core.infra.web.config.swagger.SwaggerAuthenticationStrategies.BEARER_AUTHENTICATION;
-import static com.lsadf.core.infra.web.config.swagger.SwaggerAuthenticationStrategies.OAUTH2_AUTHENTICATION;
-import static com.lsadf.core.infra.web.controller.ParameterConstants.GAME_SAVE_ID;
-
-import com.lsadf.application.controller.constant.ApiPathConstants;
-import com.lsadf.application.controller.constant.SwaggerConstants;
-import com.lsadf.core.infra.web.dto.request.game.characteristics.CharacteristicsRequest;
-import com.lsadf.core.infra.web.dto.response.ApiResponse;
-import com.lsadf.core.infra.web.dto.response.game.save.characteristics.CharacteristicsResponse;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
-import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.validation.Valid;
-import java.util.UUID;
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.oauth2.jwt.Jwt;
-import org.springframework.web.bind.annotation.*;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-/** Controller for characteristics operations */
-@RequestMapping(value = ApiPathConstants.CHARACTERISTICS)
-@Tag(name = SwaggerConstants.CHARACTERISTICS_CONTROLLER)
-@SecurityRequirement(name = BEARER_AUTHENTICATION)
-@SecurityRequirement(name = OAUTH2_AUTHENTICATION)
-public interface CharacteristicsController {
-  /**
-   * Updates the characteristics of a game save
-   *
-   * @param jwt Jwt
-   * @param gameSaveId the game save id
-   * @param characteristicsRequest the characteristics request
-   * @return the characteristics
-   */
-  @PostMapping(value = Constants.ApiPaths.GAME_SAVE_ID)
-  @Operation(summary = "Updates the characteristics of a game save")
-  @ApiResponses(
-      value = {
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(
-            responseCode = "200",
-            description = "Characteristics updated"),
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(
-            responseCode = "401",
-            description = "Unauthorized"),
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(
-            responseCode = "403",
-            description = "Forbidden"),
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(
-            responseCode = "404",
-            description = "Not found"),
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(
-            responseCode = "500",
-            description = "Internal server error")
-      })
-  ResponseEntity<ApiResponse<Void>> saveCharacteristics(
-      @AuthenticationPrincipal Jwt jwt,
-      @PathVariable(value = GAME_SAVE_ID) UUID gameSaveId,
-      @RequestBody @Valid CharacteristicsRequest characteristicsRequest);
-
-  /**
-   * Gets the characteristics of a game save
-   *
-   * @param jwt Jwt
-   * @param gameSaveId the game save id
-   * @return the characteristics
-   */
-  @GetMapping(value = Constants.ApiPaths.GAME_SAVE_ID)
-  @Operation(summary = "Gets the characteristics of a game save")
-  @ApiResponses(
-      value = {
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(
-            responseCode = "401",
-            description = "Unauthorized"),
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(
-            responseCode = "403",
-            description = "Forbidden"),
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(
-            responseCode = "200",
-            description = "OK"),
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(
-            responseCode = "404",
-            description = "Not Found"),
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(
-            responseCode = "500",
-            description = "Internal Server Error")
-      })
-  ResponseEntity<ApiResponse<CharacteristicsResponse>> getCharacteristics(
-      @AuthenticationPrincipal Jwt jwt, @PathVariable(value = GAME_SAVE_ID) UUID gameSaveId);
-
-  @NoArgsConstructor(access = AccessLevel.PRIVATE)
-  class Constants {
-    @NoArgsConstructor(access = AccessLevel.PRIVATE)
-    public static final class ApiPaths {
-      public static final String GAME_SAVE_ID = "/{game_save_id}";
-    }
-  }
-}
