+++ lsadf_api/src/main/java/com/lsadf/application/controller/game/save/characteristics/CharacteristicsControllerImpl.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.lsadf.application.controller.game.characteristics;
+package com.lsadf.application.controller.game.save.characteristics;
 
 import static com.lsadf.core.infra.web.config.auth.TokenUtils.getUsernameFromJwt;
 import static com.lsadf.core.infra.web.dto.response.ResponseUtils.generateResponse;
