+++ lsadf_api/src/main/java/com/lsadf/application/controller/game/save/currency/CurrencyController.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.lsadf.application.controller.game.game_save;
+package com.lsadf.application.controller.game.save.currency;
 
 import static com.lsadf.core.infra.web.config.swagger.SwaggerAuthenticationStrategies.BEARER_AUTHENTICATION;
 import static com.lsadf.core.infra.web.config.swagger.SwaggerAuthenticationStrategies.OAUTH2_AUTHENTICATION;
@@ -21,16 +21,14 @@ import static com.lsadf.core.infra.web.controller.ParameterConstants.GAME_SAVE_I
 
 import com.lsadf.application.controller.constant.ApiPathConstants;
 import com.lsadf.application.controller.constant.SwaggerConstants;
-import com.lsadf.core.infra.web.dto.request.game.save.update.GameSaveNicknameUpdateRequest;
+import com.lsadf.core.infra.web.dto.request.game.currency.CurrencyRequest;
 import com.lsadf.core.infra.web.dto.response.ApiResponse;
-import com.lsadf.core.infra.web.dto.response.ResponseMessages;
-import com.lsadf.core.infra.web.dto.response.game.save.GameSaveResponse;
+import com.lsadf.core.infra.web.dto.response.game.save.currency.CurrencyResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
-import java.util.List;
 import java.util.UUID;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
@@ -39,101 +37,69 @@ import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.web.bind.annotation.*;
 
-/** Controller for game save operations */
-@RequestMapping(value = ApiPathConstants.GAME_SAVE)
-@Tag(name = SwaggerConstants.GAME_SAVE_CONTROLLER)
+/** Controller for currency related operations. */
+@RequestMapping(value = ApiPathConstants.CURRENCY)
+@Tag(name = SwaggerConstants.CURRENCY_CONTROLLER)
 @SecurityRequirement(name = BEARER_AUTHENTICATION)
 @SecurityRequirement(name = OAUTH2_AUTHENTICATION)
-public interface GameSaveController {
+public interface CurrencyController {
 
-  /**
-   * Generates a new game, returns the generated game save
-   *
-   * @return the generated game save
-   */
-  @PostMapping(value = Constants.ApiPaths.GENERATE)
-  @Operation(summary = "Generates a new game, returns the generated game save")
+  @PostMapping(value = Constants.ApiPaths.GAME_SAVE_ID)
+  @Operation(summary = "Saves one or several currency amounts for a game save")
   @ApiResponses(
       value = {
         @io.swagger.v3.oas.annotations.responses.ApiResponse(
             responseCode = "401",
-            description = ResponseMessages.UNAUTHORIZED),
+            description = "Unauthorized"),
+        @io.swagger.v3.oas.annotations.responses.ApiResponse(
+            responseCode = "400",
+            description = "Bad Request"),
         @io.swagger.v3.oas.annotations.responses.ApiResponse(
             responseCode = "403",
-            description = ResponseMessages.FORBIDDEN),
+            description = "Forbidden"),
         @io.swagger.v3.oas.annotations.responses.ApiResponse(
             responseCode = "200",
-            description = ResponseMessages.OK),
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(
-            responseCode = "500",
-            description = ResponseMessages.INTERNAL_SERVER_ERROR)
-      })
-  ResponseEntity<ApiResponse<GameSaveResponse>> generateNewGameSave(
-      @AuthenticationPrincipal Jwt jwt);
-
-  /**
-   * Updates the nickname of a game save
-   *
-   * @param jwt Jwt
-   * @param id id of the game save
-   * @param gameSaveNicknameUpdateRequest GameSaveNicknameUpdateRequest
-   * @return ApiResponse
-   */
-  @PostMapping(value = Constants.ApiPaths.UPDATE_NICKNAME)
-  @Operation(summary = "Updates the nickname of a game save")
-  @ApiResponses(
-      value = {
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(
-            responseCode = "401",
-            description = ResponseMessages.UNAUTHORIZED),
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(
-            responseCode = "403",
-            description = ResponseMessages.FORBIDDEN),
-        @io.swagger.v3.oas.annotations.responses.ApiResponse(
-            responseCode = "200",
-            description = ResponseMessages.OK),
+            description = "OK"),
         @io.swagger.v3.oas.annotations.responses.ApiResponse(
             responseCode = "404",
-            description = ResponseMessages.NOT_FOUND),
+            description = "Not Found"),
         @io.swagger.v3.oas.annotations.responses.ApiResponse(
             responseCode = "500",
-            description = ResponseMessages.INTERNAL_SERVER_ERROR)
+            description = "Internal Server Error")
       })
-  ResponseEntity<ApiResponse<Void>> updateNickname(
+  ResponseEntity<ApiResponse<Void>> saveCurrency(
       @AuthenticationPrincipal Jwt jwt,
-      @PathVariable(value = GAME_SAVE_ID) UUID id,
-      @Valid @RequestBody GameSaveNicknameUpdateRequest gameSaveNicknameUpdateRequest);
+      @PathVariable(value = GAME_SAVE_ID) UUID gameSaveId,
+      @RequestBody @Valid CurrencyRequest currencyRequest);
 
-  /** Gets the user game saves */
-  @GetMapping(value = Constants.ApiPaths.ME)
-  @Operation(summary = "Gets the game saves of the logged user")
+  @GetMapping(value = Constants.ApiPaths.GAME_SAVE_ID)
+  @Operation(summary = "Gets the currency amounts for a game save")
   @ApiResponses(
       value = {
         @io.swagger.v3.oas.annotations.responses.ApiResponse(
             responseCode = "401",
-            description = ResponseMessages.UNAUTHORIZED),
+            description = "Unauthorized"),
         @io.swagger.v3.oas.annotations.responses.ApiResponse(
             responseCode = "403",
-            description = ResponseMessages.FORBIDDEN),
+            description = "Forbidden"),
         @io.swagger.v3.oas.annotations.responses.ApiResponse(
             responseCode = "200",
-            description = ResponseMessages.OK),
+            description = "OK"),
         @io.swagger.v3.oas.annotations.responses.ApiResponse(
             responseCode = "404",
-            description = ResponseMessages.NOT_FOUND),
+            description = "Not Found"),
         @io.swagger.v3.oas.annotations.responses.ApiResponse(
             responseCode = "500",
-            description = ResponseMessages.INTERNAL_SERVER_ERROR)
+            description = "Internal Server Error")
       })
-  ResponseEntity<ApiResponse<List<GameSaveResponse>>> getUserGameSaves(Jwt jwt);
+  ResponseEntity<ApiResponse<CurrencyResponse>> getCurrency(
+      @AuthenticationPrincipal Jwt jwt, @PathVariable(value = GAME_SAVE_ID) UUID gameSaveId);
 
   @NoArgsConstructor(access = AccessLevel.PRIVATE)
   class Constants {
     @NoArgsConstructor(access = AccessLevel.PRIVATE)
     public static final class ApiPaths {
-      public static final String GENERATE = "/generate";
-      public static final String ME = "/me";
-      public static final String UPDATE_NICKNAME = "/{game_save_id}/nickname";
+      public static final String GAME_SAVE_ID = "/{game_save_id}";
     }
   }
 }
